+++ mobile/android/android-components/components/feature/downloads/src/test/java/mozilla/components/feature/downloads/manager/FetchDownloadManagerTest.kt
@@ -20,6 +20,8 @@ import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.fetch.Client
 import mozilla.components.feature.downloads.AbstractFetchDownloadService
 import mozilla.components.feature.downloads.AbstractFetchDownloadService.Companion.EXTRA_DOWNLOAD_STATUS
+import mozilla.components.feature.downloads.FileSizeFormatter
+import mozilla.components.feature.downloads.fake.FakeFileSizeFormatter
 import mozilla.components.support.base.android.NotificationsDelegate
 import mozilla.components.support.test.any
 import mozilla.components.support.test.libstate.ext.waitUntilIdle
@@ -318,5 +320,6 @@ class FetchDownloadManagerTest {
         override val httpClient: Client = mock()
         override val store: BrowserStore = mock()
         override val notificationsDelegate: NotificationsDelegate = mock()
+        override val fileSizeFormatter: FileSizeFormatter = FakeFileSizeFormatter()
     }
 }