+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/downloads/DownloadService.kt
@@ -6,6 +6,7 @@ package org.mozilla.samples.browser.downloads
 
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.feature.downloads.AbstractFetchDownloadService
+import mozilla.components.feature.downloads.FileSizeFormatter
 import mozilla.components.support.base.android.NotificationsDelegate
 import org.mozilla.samples.browser.ext.components
 
@@ -13,4 +14,5 @@ class DownloadService : AbstractFetchDownloadService() {
     override val httpClient by lazy { components.client }
     override val store: BrowserStore by lazy { components.store }
     override val notificationsDelegate: NotificationsDelegate by lazy { components.notificationsDelegate }
+    override val fileSizeFormatter: FileSizeFormatter by lazy { components.fileSizeFormatter }
 }