+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/downloads/DownloadService.kt
@@ -6,6 +6,7 @@ package org.mozilla.fenix.downloads
 
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.feature.downloads.AbstractFetchDownloadService
+import mozilla.components.feature.downloads.FileSizeFormatter
 import mozilla.components.support.base.android.NotificationsDelegate
 import org.mozilla.fenix.R
 import org.mozilla.fenix.ext.components
@@ -15,4 +16,5 @@ class DownloadService : AbstractFetchDownloadService() {
     override val store: BrowserStore by lazy { components.core.store }
     override val style: Style by lazy { Style(R.color.fx_mobile_text_color_accent) }
     override val notificationsDelegate: NotificationsDelegate by lazy { components.notificationsDelegate }
+    override val fileSizeFormatter: FileSizeFormatter by lazy { components.core.fileSizeFormatter }
 }