+++ mobile/android/focus-android/app/src/main/java/org/mozilla/focus/downloads/DownloadService.kt
@@ -7,6 +7,7 @@ package org.mozilla.focus.downloads
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.fetch.Client
 import mozilla.components.feature.downloads.AbstractFetchDownloadService
+import mozilla.components.feature.downloads.FileSizeFormatter
 import mozilla.components.support.base.android.NotificationsDelegate
 import org.mozilla.focus.ext.components
 
@@ -14,4 +15,5 @@ class DownloadService : AbstractFetchDownloadService() {
     override val httpClient: Client by lazy { components.client }
     override val store: BrowserStore by lazy { components.store }
     override val notificationsDelegate: NotificationsDelegate by lazy { components.notificationsDelegate }
+    override val fileSizeFormatter: FileSizeFormatter by lazy { components.fileSizeFormatter }
 }