+++ mobile/android/android-components/docs/changelog.md
@@ -5,6 +5,19 @@ permalink: /changelog/
 ---
 
 # 139.0 (In Development)
+* **feature-downloads**
+  * ⚠️ **Breaking change**: Added new `fileSizeFormatter` abstract val to `AbstractFetchDownloadService`. [Bug 1956580](https://bugzilla.mozilla.org/show_bug.cgi?id=1956580).
+  * ⚠️ **Breaking change**: Added new `fileSizeFormatter` parameter to `DownloadNotification.createDownloadGroupNotification`. [Bug 1956580](https://bugzilla.mozilla.org/show_bug.cgi?id=1956580).
+    * Changed download notification group to display the size downloaded so far and the total file size, instead of the download percentage completed.
+  * ⚠️ **Breaking change**: Added new `fileSizeFormatter` parameter to `DownloadNotification.createOngoingDownloadNotification`. [Bug 1956580](https://bugzilla.mozilla.org/show_bug.cgi?id=1956580).
+    * Removed the download percentage completed in the notification summary.
+    * Changed the content text to display the size downloaded so far and the total file size, instead of the download percentage completed.
+  * ⚠️ **Breaking change**: Added new `fileSizeFormatter` parameter to `DownloadNotification.getSummaryList`. [Bug 1956580](https://bugzilla.mozilla.org/show_bug.cgi?id=1956580).
+    * Changed download notification group to display the size downloaded so far and the total file size, instead of the download percentage completed.
+  * ⚠️ **Breaking change**: Added new `fileSizeFormatter` parameter to `DownloadNotification.getProgress`. [Bug 1956580](https://bugzilla.mozilla.org/show_bug.cgi?id=1956580).
+    * Changed the function to return the size downloaded so far and the total file size, instead of the download percentage completed.
+  * ⚠️ **Breaking change**: Added new `fileSizeFormatter` parameter to `DownloadNotification.getStatusDescription`. [Bug 1956580](https://bugzilla.mozilla.org/show_bug.cgi?id=1956580).
+    * Changed the return of `DOWNLOADING` status to display the size downloaded so far and the total file size, instead of the download percentage completed.
 
 # 138.0
 * **support-ktx**